F10 Phase 3: split webviewHtml.ts state/util helpers into modules
Extract escapeHtml, getNonce, jsonForScript into htmlUtils.ts and getMrdStateHtml/getMrdLoadingHtml/getMrdErrorHtml into stateHtml.ts. webviewHtml.ts keeps getMrdViewerHtml and getMrdBackendMissingHtml. No CSP, localResourceRoots, or build changes.
